Move alias.js to alias/index.js
